+++ source/MarketRoles.B2B.CimMessageAdapter.IntegrationTests/MessageReceiverTests.cs
@@ -14,7 +14,7 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
     public class MessageReceiverTests
     {
         private readonly MessageIdStore _messageIdStore = new();
-        private ActivityRecordForwarder _activityRecordForwarder;
+        private ActivityRecordForwarderStub _activityRecordForwarderStub;
         private readonly TransactionIdsStub _transactionIdsStub = new();
 
         public MessageReceiverTests()
@@ -69,7 +69,7 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
             await ReceiveRequestChangeOfSupplierMessage(CreateMessageNotConformingToXmlSchema())
                 .ConfigureAwait(false);
 
-            var activityRecord = _activityRecordForwarder.CommittedItems.FirstOrDefault();
+            var activityRecord = _activityRecordForwarderStub.CommittedItems.FirstOrDefault();
             Assert.NotNull(activityRecord);
             Assert.Equal("12345699", activityRecord.MRid);
             Assert.Equal("579999993331812345", activityRecord.MarketEvaluationPointmRID);
@@ -89,7 +89,7 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
             await ReceiveRequestChangeOfSupplierMessage(CreateMessage())
                 .ConfigureAwait(false);
 
-            Assert.Empty(_activityRecordForwarder.CommittedItems);
+            Assert.Empty(_activityRecordForwarderStub.CommittedItems);
         }
 
         [Fact]
@@ -98,7 +98,7 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
             await ReceiveRequestChangeOfSupplierMessage(CreateMessageWithDuplicateTransactionIds())
                 .ConfigureAwait(false);
 
-            Assert.Single(_activityRecordForwarder.CommittedItems);
+            Assert.Single(_activityRecordForwarderStub.CommittedItems);
         }
 
 
@@ -110,8 +110,8 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
 
         private MessageReceiver CreateMessageReceiver()
         {
-            _activityRecordForwarder = new ActivityRecordForwarder();
-            var messageReceiver = new MessageReceiver(_messageIdStore, _activityRecordForwarder, _transactionIdsStub);
+            _activityRecordForwarderStub = new ActivityRecordForwarderStub();
+            var messageReceiver = new MessageReceiver(_messageIdStore, _activityRecordForwarderStub, _transactionIdsStub);
             return messageReceiver;
         }
 
@@ -150,38 +150,17 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
         }
     }
 
-    public class ActivityRecordForwarder : IActivityRecordForwarder
-    {
-        private readonly List<MarketActivityRecord> _uncommittedItems = new();
-        private readonly List<MarketActivityRecord> _committedItems = new();
-        public IReadOnlyCollection<MarketActivityRecord> CommittedItems => _committedItems.AsReadOnly();
-
-        public async Task AddAsync(MarketActivityRecord marketActivityRecord)
-        {
-            _committedItems.Clear();
-            _uncommittedItems.Add(marketActivityRecord);
-        }
-
-        public Task CommitAsync()
-        {
-            _committedItems.Clear();
-            _committedItems.AddRange(_uncommittedItems);
-            _uncommittedItems.Clear();
-            return Task.CompletedTask;
-        }
-    }
-
     public class MessageReceiver
     {
         private readonly List<Error> _errors = new();
         private readonly MessageIdStore _messageIds;
-        private readonly ActivityRecordForwarder _activityRecordForwarder;
+        private readonly ActivityRecordForwarderStub _activityRecordForwarderStub;
         private readonly TransactionIdsStub _transactionIdsStub;
 
-        public MessageReceiver(MessageIdStore messageIds, ActivityRecordForwarder activityRecordForwarder, TransactionIdsStub transactionIdsStub)
+        public MessageReceiver(MessageIdStore messageIds, ActivityRecordForwarderStub activityRecordForwarderStub, TransactionIdsStub transactionIdsStub)
         {
             _messageIds = messageIds ?? throw new ArgumentNullException(nameof(messageIds));
-            _activityRecordForwarder = activityRecordForwarder ?? throw new ArgumentNullException(nameof(activityRecordForwarder));
+            _activityRecordForwarderStub = activityRecordForwarderStub ?? throw new ArgumentNullException(nameof(activityRecordForwarderStub));
             _transactionIdsStub = transactionIdsStub;
         }
 
@@ -327,7 +306,7 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
 
             if (hasInvalidHeaderValues == false)
             {
-                await _activityRecordForwarder.CommitAsync().ConfigureAwait(false);
+                await _activityRecordForwarderStub.CommitAsync().ConfigureAwait(false);
             }
             return _errors.Count == 0 ? Result.Succeeded() : Result.Failure(_errors.ToArray());
         }
@@ -340,7 +319,7 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
 
         private Task StoreActivityRecordAsync(MarketActivityRecord marketActivityRecord)
         {
-            return _activityRecordForwarder.AddAsync(marketActivityRecord);
+            return _activityRecordForwarderStub.AddAsync(marketActivityRecord);
         }
 
         private Task<bool> CheckMessageIdAsync(string messageId)
